Rename dct-kit to dct_kit
