+++ deepcell_types/dct_kit/config/celltype_embeddings-text-embedding-3-large-1024.json
